feat: add refresh button to index page
