+++ renderer/src/routes/index.tsx
@@ -4,6 +4,7 @@ import {
   getApiV1BetaWorkloadsQueryKey,
   getApiV1BetaWorkloadsByNameOptions,
 } from '@/common/api/generated/@tanstack/react-query.gen'
+import { RefreshButton } from '@/common/components/refresh-button'
 import { useToastMutation } from '@/common/hooks/use-toast-mutation'
 import { pollServerStatus } from '@/common/lib/polling'
 import { DialogFormRunMcpServerWithCommand } from '@/features/mcp-servers/components/dialog-form-run-mcp-command'
@@ -22,7 +23,7 @@ export const Route = createFileRoute('/')({
 })
 
 export function Index() {
-  const { data } = useSuspenseQuery({
+  const { data, refetch } = useSuspenseQuery({
     ...getApiV1BetaWorkloadsOptions({ query: { all: true } }),
   })
   const workloads = data?.workloads ?? []
@@ -34,10 +35,12 @@ export function Index() {
     <>
       <div className="mb-6 flex items-center">
         <h1 className="text-3xl font-semibold">Installed</h1>
-        <DropdownMenuRunMcpServer
-          openRunCommandDialog={() => setIsRunWithCommandOpen(true)}
-          className="ml-auto"
-        />
+        <div className="ml-auto flex gap-2">
+          <RefreshButton refresh={refetch} />
+          <DropdownMenuRunMcpServer
+            openRunCommandDialog={() => setIsRunWithCommandOpen(true)}
+          />
+        </div>
         <DialogFormRunMcpServerWithCommand
           isOpen={isRunWithCommandOpen}
           onOpenChange={setIsRunWithCommandOpen}
